// src/core/models/packages/packageAttribute.ts
export interface PackageAttribute {
  activeText: string | null;
  actualValue: number | null;
  amount: number | null;
  attributeID: number;
  attributeName: string;
  isActive: boolean | null;
  packageDetailsID: number;
  packageID: number;
  pD_ActiveText: string | null;
  pD_isActive: boolean | null;
  sortOrder: number | null;
  value: string | null;
}